refactor: 제시된 요구사항의 따른 자료형 리팩토링(Long -> Integer형)

+++ src/main/java/org/prography/spring/domain/User.java
@@ -19,11 +19,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(unique = true)
-    private Long fakerId;
+    private Integer fakerId;
 
     @NotNull
     @Column(length = 255)
@@ -46,7 +46,7 @@ public class User {
 
     @Builder
     private User(
-            Long fakerId,
+            Integer fakerId,
             String name,
             String email,
             UserStatus status
